refactor(docs): move the route provide and the family effect into `Theme.setup`
- Call `provideCurrentRule` and `useFamilyDataset` from the theme's `setup`, which the alpha un-deprecated
- Extract the body `data-family` effect into `useFamilyDataset`, seating it under `lib/composables`
- Reduce `Layout.vue` to its slot overrides, leaving no logic in its script block
- Cover the new composable against a rule route and a route outside the rules tree
