+++ site/.vitepress/theme/Layout.vue
@@ -1,8 +1,5 @@
 <script setup lang="ts">
-import DefaultTheme    from 'vitepress/theme'
-import { watchEffect } from 'vue'
-
-import { provideCurrentRule, useCurrentFamily } from '../lib/composables/route'
+import DefaultTheme from 'vitepress/theme'
 
 import BuildMetadata      from './components/layout/BuildMetadata.vue'
 import GlossaryFolioIndex from './components/glossary/GlossaryFolioIndex.vue'
@@ -10,14 +7,6 @@ import NotFound           from './components/layout/NotFound.vue'
 import RuleChrome         from './components/rules/RuleChrome.vue'
 import RuleLinkPopper     from './components/rules/RuleLinkPopper.vue'
 import StarBadge          from './components/layout/StarBadge.vue'
-
-provideCurrentRule()
-const family = useCurrentFamily()
-watchEffect(() => {
-  if (typeof document === 'undefined') return
-  if (family.value) document.body.dataset.family = family.value
-  else              delete document.body.dataset.family
-})
 </script>
 
 <template>
